Create c10 library for shared functionality between PyTorch and Caffe2.
The c10 library is used by both PyTorch and Caffe2.  We start it off
small and minimal, and we will suck in functionality from PyTorch/Caffe2
that we want to use from the other side.

To start, c10 contains:

c10/Optional.h    An implementation of std::optional
c10/Error.h       Error reporting functionality
c10/Registry.h    The Caffe2 registry
