+++ aten/src/ATen/CMakeLists.txt
@@ -242,7 +242,8 @@ ENDIF()
 # of NO_CUDA (it always builds the same way); libATen_cuda.so is only
 # built when NO_CUDA=0 and CUDA is available.
 
-add_library(ATen_cpu ${AT_LINK_STYLE} ${all_cpu_cpp})
+add_library(ATen_cpu ${AT_LINK_STYLE} ${all_cpu_cpp} ${C10_SRCS})
+target_link_libraries(ATen_cpu)
 add_dependencies(ATen_cpu aten_files_are_generated)
 if(NOT NO_CUDA)
   cuda_add_library(ATen_cuda ${AT_LINK_STYLE} ${all_cuda_cpp})
@@ -512,6 +513,8 @@ if(NOT NO_CUDA)
   get_filename_component(ATEN_CUDA_OUTPUT_NAME ${ATEN_CUDA_OUTPUT_NAME} NAME)
 endif()
 
+# TODO: Technically, C10 does not respect AT_INSTALL_LIB_DIR at the moment.
+# But this entire business seems highly questionable
 set(ATEN_LIBRARIES
   "${CMAKE_INSTALL_PREFIX}/${AT_INSTALL_LIB_DIR}/${ATEN_CPU_OUTPUT_NAME}")
 
@@ -538,6 +541,10 @@ ENDFOREACH()
 FOREACH(HEADER ${cuda_h})
   GET_FILENAME_COMPONENT(DIR ${HEADER} DIRECTORY)
   INSTALL(FILES ${HEADER} DESTINATION ${AT_INSTALL_INCLUDE_DIR}/ATen/${DIR})
+ENDFOREACH()
+FOREACH(HEADER ${C10_HEADERS})
+  GET_FILENAME_COMPONENT(DIR ${HEADER} DIRECTORY)
+  INSTALL(FILES ${HEADER} DESTINATION ${AT_INSTALL_INCLUDE_DIR}/c10/${DIR})
 ENDFOREACH()
 INSTALL(FILES ${CMAKE_CURRENT_BINARY_DIR}/ATen/Declarations.yaml
   DESTINATION ${AT_INSTALL_SHARE_DIR}/ATen)
