+++ aten/src/ATen/detail/CUDAHooksInterface.cpp
@@ -23,6 +23,6 @@ namespace detail {
   }
 }
 
-AT_DEFINE_REGISTRY(CUDAHooksRegistry, CUDAHooksInterface, CUDAHooksArgs)
+C10_DEFINE_REGISTRY(CUDAHooksRegistry, CUDAHooksInterface, CUDAHooksArgs)
 
 } // namespace at
