+++ aten/src/ATen/detail/VariableHooksInterface.cpp
@@ -24,6 +24,6 @@ namespace detail {
 
 }
 
-AT_DEFINE_REGISTRY(VariableHooksRegistry, VariableHooksInterface, VariableHooksArgs)
+C10_DEFINE_REGISTRY(VariableHooksRegistry, VariableHooksInterface, VariableHooksArgs)
 
 } // namespace at::detail
