+++ caffe2/core/blob_serialization.h
@@ -24,13 +24,14 @@ constexpr auto kTensorBlobType = "Tensor";
 constexpr auto kChunkIdSeparator = "#%";
 
 // The Blob serialization registry and serializer creator functions.
-CAFFE_DECLARE_TYPED_REGISTRY(
+C10_DECLARE_TYPED_REGISTRY_NOARG(
+    CAFFE2_API,
     BlobSerializerRegistry,
     CaffeTypeId,
     BlobSerializerBase,
     std::unique_ptr);
 #define REGISTER_BLOB_SERIALIZER(id, ...) \
-  CAFFE_REGISTER_TYPED_CLASS(BlobSerializerRegistry, id, __VA_ARGS__)
+  C10_REGISTER_TYPED_CLASS(BlobSerializerRegistry, id, __VA_ARGS__)
 // Creates an operator with the given operator definition.
 inline unique_ptr<BlobSerializerBase> CreateSerializer(CaffeTypeId id) {
   return BlobSerializerRegistry()->Create(id);
@@ -82,9 +83,9 @@ class BlobDeserializerBase {
   virtual void Deserialize(const BlobProto& proto, Blob* blob) = 0;
 };
 
-CAFFE_DECLARE_REGISTRY(BlobDeserializerRegistry, BlobDeserializerBase);
+C10_DECLARE_REGISTRY_NOARG(CAFFE2_API, BlobDeserializerRegistry, BlobDeserializerBase);
 #define REGISTER_BLOB_DESERIALIZER(name, ...) \
-  CAFFE_REGISTER_CLASS(BlobDeserializerRegistry, name, __VA_ARGS__)
+  C10_REGISTER_CLASS(BlobDeserializerRegistry, name, __VA_ARGS__)
 // Creates an operator with the given operator definition.
 inline unique_ptr<BlobDeserializerBase> CreateDeserializer(const string& type) {
   return BlobDeserializerRegistry()->Create(type);
