+++ caffe2/core/transform.cc
@@ -10,7 +10,7 @@ namespace caffe2 {
 
 using transform::Graph;
 
-CAFFE_DEFINE_REGISTRY(TransformRegistry, Transform);
+C10_DEFINE_REGISTRY_NOARG(TransformRegistry, Transform);
 
 std::vector<std::vector<int>> Transform::PatternMatch(const Graph& graph) {
   // checks if the node at index i is matched already or not
