+++ caffe2/core/transform.h
@@ -150,9 +150,9 @@ class Transform {
 // Creates a Transform based on a key, which should be defined in registry.
 unique_ptr<Transform> CreateTransform(string key);
 
-CAFFE_DECLARE_REGISTRY(TransformRegistry, Transform);
+C10_DECLARE_REGISTRY_NOARG(CAFFE2_API, TransformRegistry, Transform);
 #define REGISTER_TRANSFORM(name, ...) \
-  CAFFE_REGISTER_CLASS(TransformRegistry, name, __VA_ARGS__)
+  C10_REGISTER_CLASS(TransformRegistry, name, __VA_ARGS__)
 
 // Create a Transform object from registry,
 // and immediately apply it to a Netdef.
