+++ caffe2/mobile/contrib/arm-compute/core/operator.cc
@@ -2,7 +2,7 @@
 
 namespace caffe2 {
 
-CAFFE_DEFINE_REGISTRY(GLOperatorRegistry, OperatorBase, const OperatorDef &,
+C10_DEFINE_REGISTRY(GLOperatorRegistry, OperatorBase, const OperatorDef &,
                       Workspace *);
 CAFFE_REGISTER_DEVICE_TYPE(DeviceType::OPENGL, GLOperatorRegistry);
